+++ generalimport/__init__.py
@@ -1,12 +1,8 @@
 
-from generalimport.generalimport_bottom import get_installed_modules_names, module_is_installed, import_module, module_name_is_namespace, module_is_namespace, spec_is_namespace, _get_previous_frame_filename, _get_top_name, _get_scope_from_filename, get_spec, fake_module_check
+from generalimport.generalimport_bottom import get_installed_modules_names, module_is_installed, import_module, module_name_is_namespace, module_is_namespace, spec_is_namespace, get_spec, fake_module_check
 from generalimport.exception import MissingDependencyException, MissingOptionalDependency
-from generalimport.fake_module import FakeModule, is_imported
+from generalimport.dunders import DynamicDunder, NON_CALLABLE_DUNDERS, CALLABLE_CLASS_DUNDERS, CALLABLE_DUNDERS
+from generalimport.fake_module import FakeModule
 from generalimport.general_importer import GeneralImporter
 from generalimport.import_catcher import ImportCatcher
-from generalimport.top import generalimport, get_importer, reset_generalimport
-
-
-
-
-
+from generalimport.top import generalimport, get_importer, reset_generalimport, is_imported
